More work on 2nd navbar
The navbar is looking a bit better still.  I've added content to the
courthouse page just to test the fixing of the navbar.  This will be
removed when the navbar remains static, which is not happening yet.
Internet down here so I can't look at the bootstrap api for reference at
the moment.