More work on slideshow
The slideshow is working a bit better now.  It is scrolling through the
images automatically, but there are issues.
The first image is much larger than the others, and the other images are
also slightly different sizes, so the webpage "jumps" a bit to
accommodate the changes.  Also, the previous and next arrows do not
work.