Fixed slideshow images
The images in the slideshow are now are the same size.  It wouldn't work
for me for ages, and I really don't know what I did differently this
time, but it worked!

I also inserted the 5 sites into the courthouse webpage.  I think it
will be better to have them all on the same page, rather than a page
each.