Rearranged the sites
The sites were fine on a large screen, but when I checked it on a
smaller screen, things didn't look right.  ie. - It appeared, text
image, image text, text image, etc.  I wanted it to be text image all
the way down, for usability and uniformity.  It was done using push and
pull on the columns.