Added Italian
Added Italian text, along with a button for that language option.
Also, there was an issue where when the user clicked on a language
button once, it was fine, but if they clicked a second time in a row,
the text faded away and no text was to be seen.  This is now fixed, but
there is no effect working when the text changes from one language to
another.